requirements.txt
pyglet<=1.5.0,>=1.4.0
gym>=0.9.0
numpy>=1.10.0

pyzmq>=16.0.0
scikit-image>=0.13.1
opencv-python>=3.4
pyyaml>=3.11
cloudpickle>=1.2.0,<1.4.0

dataclasses<=0.6; python_version < '3.7'
carnivalmirror==0.6.2
duckietown-world-daffy

# testing
pipdeptree

torch
Pillow>=8.0.1
